requirements.txt
streamlit==1.31.1
pydrive2==1.19.0
google-api-python-client==2.119.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
python-dotenv==1.0.1
llama-index==0.10.12
llama-index-readers-google==0.1.4
llama-index-embeddings-huggingface==0.1.3
docx2txt==0.8
torch==2.2.1
transformers==4.38.1
python-pptx==0.6.23
pillow==10.2.0
llama-index-embeddings-instructor==0.1.3
instructorembedding==1.0.1
sentence-transformers==2.2.2
llama-index-llms-palm==0.1.3
google-generativeai==0.3.2